# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Leedz API
  description: |
    API for Leedz - Sales Intelligence & CRM Platform
    
    ## Authentication
    All endpoints (except public ones) require a Bearer token in the Authorization header.
    
    ## Tenant Context
    The current tenant is determined from the JWT token. Use `/api/auth/switch-tenant` to change tenants.
    
    ## Rate Limiting
    - Default: 100 requests/minute
    - Search: 10 requests/minute
    - WhatsApp: 100 requests/hour
  version: 2.0.0
  contact:
    name: Leedz Team
    email: dev@leadz.sa
servers:
  - url: https://api.leadz.sa
    description: Production
  - url: https://api.staging.leadz.sa
    description: Staging
  - url: http://localhost:3000
    description: Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
    UserRole:
      type: string
      enum: [OWNER, ADMIN, MANAGER, SALES]
...
    JobType:
      type: string
      enum: [SEARCH, SURVEY, WHATSAPP, WHATSAPP_BULK, IMPORT, EXPORT, REVEAL, REPORT, SYNC]
    JobStatus:
      type: string
      enum: [PENDING, RUNNING, SUCCESS, FAILED, CANCELLED]
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/JobType'
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
          nullable: true
        entityType:
          type: string
          nullable: true
        entityId:
          type: string
          format: uuid
          nullable: true
        result:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
          properties:
            code:
              type: string
            message:
              type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true

    # ==================== Evidence & Reports ====================
...
    TeamMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          type: string
          enum: [ACTIVE, SUSPENDED]
        avatarUrl:
          type: string
          nullable: true
        joinedAt:
          type: string
          format: date-time
...
paths:
...
  /api/jobs:
    get:
      tags: [Jobs]
      summary: List jobs
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JobStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/JobType'
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      jobs:
                        type: array
                        items:
                          $ref: '#/components/schemas/Job'
  /api/jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      job:
                        $ref: '#/components/schemas/Job'

  # ==================== WhatsApp ====================
...
  /api/team:
    get:
      tags: [Team]
      summary: List team members
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of team members
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      members:
                        type: array
                        items:
                          $ref: '#/components/schemas/TeamMember'
...
  /api/team/{id}/role:
    patch:
      tags: [Team]
      summary: Change member role
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  $ref: '#/components/schemas/UserRole'
      responses:
        '200':
          description: Role updated
  /api/team/{id}:
    delete:
      tags: [Team]
      summary: Remove team member
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Member removed

  # ==================== Billing ====================
...
  /api/tenants:
    get:
      tags: [Tenants]
      summary: List user's tenants
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      tenants:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tenant'
    post:
      tags: [Tenants]
      summary: Create new tenant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                slug:
                  type: string
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      tenant:
                        $ref: '#/components/schemas/Tenant'
                      membership:
                        type: object
                        properties:
                          role:
                            $ref: '#/components/schemas/UserRole'
  /api/tenants/{id}:
    patch:
      tags: [Tenants]
      summary: Update tenant
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                logoUrl:
                  type: string
      responses:
        '200':
          description: Tenant updated

  # ==================== Invites (CF-11 fix) ====================
...